Fix elements.css delete error

+++ deploy.py
@@ -61,6 +61,10 @@ def cleanup_remote_directory(api, deploy_dir):
     if deploy_dir == "/":
         items = api.list_files(deploy_dir)
         for item in items:
+            # skip `elements.css` as it cannot be deleted
+            if item["name"] == "elements.css":
+                continue
+
             full_path = os.path.join(deploy_dir, item["name"])
             logger.info({"message": "Deleting file", "file": full_path})
             api.delete_file_or_directory(full_path)
